Add routes and validators

+++ routers/contacts.js
@@ -1,7 +1,12 @@
-const express = require("express");
 const boom = require("@hapi/boom");
+const express = require("express");
 const router = express.Router();
+
 const { Contact, PhoneNumber, Email, Address } = require("../db");
+const validateContact = require("../validators/contactValidator");
+const { validatePhoneNumbers } = require("../validators/phoneNumberValidator");
+const { validateEmails } = require("../validators/emailValidator");
+const { validateAddresses } = require("../validators/addressValidator");
 
 // Get contacts
 router.get("/:accountId", async (req, res, next) => {
@@ -20,8 +25,8 @@ router.get("/:accountId", async (req, res, next) => {
 	}
 });
 
-// Get a contact
-router.get("/:contactId", async (req, res, next) => {
+// Get contact by ID
+router.get("/find/:contactId", async (req, res, next) => {
 	const { contactId } = req.params;
 
 	try {
@@ -45,6 +50,25 @@ router.get("/:contactId", async (req, res, next) => {
 router.post("/create", async (req, res, next) => {
 	try {
 		let { contact, phoneNumbers, emails, addresses } = req.body;
+
+		// Validation
+		const { error: contactError } = validateContact(contact);
+		const { error: phoneNumberError } = validatePhoneNumbers(phoneNumbers);
+		const { error: emailError } = validateEmails(emails);
+		const { error: addressError } = validateAddresses(addresses);
+		if (contactError) {
+			return next(boom.badRequest(contactError.details[0].message));
+		}
+		if (phoneNumberError) {
+			return next(boom.badRequest(phoneNumberError.details[0].message));
+		}
+		if (emailError) {
+			return next(boom.badRequest(emailError.details[0].message));
+		}
+		if (addressError) {
+			return next(boom.badRequest(addressError.details[0].message));
+		}
+
 		let result = {};
 
 		result.contact = await Contact.create(contact);
@@ -99,6 +123,12 @@ router.delete("/delete/:contactId", async (req, res, next) => {
 	const { contactId } = req.params;
 
 	try {
+		const result = await Contact.findOne({ where: { id: contactId } });
+
+		if (!result) {
+			return next(boom.badRequest("Invalid contact", { statusCode: 400 }));
+		}
+
 		await PhoneNumber.destroy({
 			where: { contactId }
 		});
@@ -108,10 +138,11 @@ router.delete("/delete/:contactId", async (req, res, next) => {
 		await Address.destroy({
 			where: { contactId }
 		});
-		// ====> Add Group as well
+
+		// ====> Delete Group
 
 		await Contact.destroy({
-			where: { contactId }
+			where: { id: contactId }
 		});
 
 		res.status(200).json({
@@ -122,177 +153,5 @@ router.delete("/delete/:contactId", async (req, res, next) => {
 		next(boom.internal(`Unable to delete contact ${contactId}`));
 	}
 });
-
-// Add phone number
-router.post("/phone-number/create/:contactId", async (req, res, next) => {
-	const { contactId } = req.params;
-
-	try {
-		const result = await PhoneNumber.create({ ...req.body, contactId });
-
-		res.json(result);
-	} catch (err) {
-		console.error(err);
-		next(boom.internal(`Unable to add phone number for contact ${contactId}`));
-		//next(boom.notImplemented("Unable to create phone number"));
-	}
-});
-
-// Update phone number
-router.patch(
-	"/phone-number/update/:contactId/:phoneId",
-	async (req, res, next) => {
-		const { contactId, phoneId } = req.params;
-
-		try {
-			await PhoneNumber.update(req.body, { where: { id: phoneId } });
-
-			res.json({
-				msg: "Phone number updated."
-			});
-		} catch (err) {
-			console.error(err);
-			next(
-				boom.internal(
-					`Unable to update phone number ${phoneId} for contact ${contactId}`
-				)
-			);
-		}
-	}
-);
-
-// Delete phone number
-router.delete(
-	"/phone-number/delete/:contactId/:phoneId",
-	async (req, res, next) => {
-		const { contactId, phoneId } = req.params;
-
-		try {
-			await PhoneNumber.destroy({ where: { id: phoneId } });
-			res.json({
-				msg: "Phone number deleted."
-			});
-		} catch (err) {
-			console.error(err);
-			next(
-				boom.internal(
-					`Unable to delete phone number ${phoneId} for contact ${contactId}`
-				)
-			);
-		}
-	}
-);
-
-// Add email
-router.post("/email/create/:contactId", async (req, res, next) => {
-	const { contactId } = req.params;
-
-	try {
-		const result = await Email.create({ ...req.body, contactId });
-
-		res.json(result);
-	} catch (err) {
-		console.error(err);
-		next(boom.internal(`Unable to add email for contact ${contactId}`));
-	}
-});
-
-// Update email
-router.patch("/email/update/:contactId/:emailId", async (req, res, next) => {
-	const { contactId, emailId } = req.params;
-
-	try {
-		await Email.update(req.body, { where: { id: emailId } });
-
-		res.json({
-			msg: "Email updated."
-		});
-	} catch (err) {
-		console.error(err);
-		next(
-			boom.internal(
-				`Unable to update email ${phoneId} for contact ${contactId}`
-			)
-		);
-	}
-});
-
-// Delete email
-router.delete("/email/delete/:contactId/:emailId", async (req, res, next) => {
-	const { contactId, emailId } = req.params;
-
-	try {
-		await Email.destroy({ where: { id: emailId } });
-		res.json({
-			msg: "Email deleted."
-		});
-	} catch (err) {
-		console.error(err);
-		next(
-			boom.internal(
-				`Unable to delete email ${emailId} for contact ${contactId}`
-			)
-		);
-	}
-});
-
-// Add address
-router.post("/address/create/:contactId", async (req, res, next) => {
-	const { contactId } = req.params;
-
-	try {
-		const result = await Address.create({ ...req.body, contactId });
-
-		res.json(result);
-	} catch (err) {
-		console.error(err);
-		next(boom.internal(`Unable to add address for contact ${contactId}`));
-	}
-});
-
-// Update address
-router.patch(
-	"/address/update/:contactId/:addressId",
-	async (req, res, next) => {
-		const { contactId, addressId } = req.params;
-
-		try {
-			await Address.update(req.body, { where: { id: addressId } });
-
-			res.json({
-				msg: "Address updated."
-			});
-		} catch (err) {
-			console.error(err);
-			next(
-				boom.internal(
-					`Unable to update address ${addressId} for contact ${contactId}`
-				)
-			);
-		}
-	}
-);
-
-// Delete address
-router.delete(
-	"/address/delete/:contactId/:addressId",
-	async (req, res, next) => {
-		const { contactId, addressId } = req.params;
-
-		try {
-			await Address.destroy({ where: { id: addressId } });
-			res.json({
-				msg: "Address deleted."
-			});
-		} catch (err) {
-			console.error(err);
-			next(
-				boom.internal(
-					`Unable to delete address ${addressId} for contact ${contactId}`
-				)
-			);
-		}
-	}
-);
 
 module.exports = router;
